+++ app/src/main/java/io/horizontalsystems/ethereumkit/sample/MainViewModel.kt
@@ -217,7 +217,7 @@ class MainViewModel : ViewModel() {
 //        ethereumAdapter.refresh()
 //        erc20Adapter.refresh()
 
-//        val storage = io.horizontalsystems.ethereumkit.core.refactoring.Storage()
+//        val storage = io.horizontalsystems.ethereumkit.refactoring.Storage()
 //        val hashPool = NotSyncedTransactionPool(storage)
 //        val syncer = TransactionSyncer(hashPool, ethereumKit.blockchain)
