+++ erc20kit/src/main/java/io/horizontalsystems/erc20kit/models/TokenBalance.kt
@@ -4,4 +4,7 @@ import androidx.room.Entity
 import java.math.BigInteger
 
 @Entity(primaryKeys = ["primaryKey"])
-class TokenBalance(val value: BigInteger, val primaryKey: String = "primaryKey")
+class TokenBalance(
+        val value: BigInteger,
+        val primaryKey: String = "primaryKey"
+)
