--- erc20kit/src/main/java/io/horizontalsystems/erc20kit/models/TokenError.kt
@@ -1,7 +1,0 @@
-package io.horizontalsystems.erc20kit.models
-
-sealed class TokenError: Exception() {
-    class InvalidAddress: TokenError()
-    class NotRegistered: TokenError()
-    class AlreadyRegistered: TokenError()
-}
