+++ erc20kit/src/main/java/io/horizontalsystems/erc20kit/models/TransactionCache.kt
@@ -5,19 +5,8 @@ import io.horizontalsystems.ethereumkit.models.Address
 import java.math.BigInteger
 import java.util.*
 
-enum class TransactionType(val value: String) {
-    TRANSFER("transfer"),
-    APPROVE("approve");
-
-    companion object {
-        fun valueOf(value: String?): TransactionType? {
-            return values().find { it.value == value }
-        }
-    }
-}
-
 @Entity(primaryKeys = ["hash", "interTransactionIndex"])
-data class TransactionRecord(
+data class TransactionCache(
         val hash: ByteArray,
         var interTransactionIndex: Int = 0,
         var logIndex: Int?,
@@ -29,7 +18,7 @@ data class TransactionRecord(
 ) {
 
     override fun equals(other: Any?): Boolean {
-        if (other !is TransactionRecord)
+        if (other !is TransactionCache)
             return false
 
         return hash.contentEquals(other.hash) && interTransactionIndex == other.interTransactionIndex
