+++ ethereumkit/src/main/java/io/horizontalsystems/ethereumkit/api/jsonrpcsubscription/RpcSubscription.kt
@@ -1,7 +1,7 @@
 package io.horizontalsystems.ethereumkit.api.jsonrpcsubscription
 
 import com.google.gson.Gson
-import io.horizontalsystems.ethereumkit.api.RpcSubscriptionResponse
+import io.horizontalsystems.ethereumkit.api.core.RpcSubscriptionResponse
 
 abstract class RpcSubscription<T>(val params: List<Any>) {
     protected abstract val typeOfResult: Class<T>
