+++ ethereumkit/src/main/java/io/horizontalsystems/ethereumkit/api/models/EtherscanResponse.kt
@@ -1,4 +1,4 @@
-package io.horizontalsystems.ethereumkit.api.models.etherscan
+package io.horizontalsystems.ethereumkit.api.models
 
 data class EtherscanResponse(
         val status: String,
