+++ ethereumkit/src/main/java/io/horizontalsystems/ethereumkit/transactionsyncers/EthereumTransactionSyncer.kt
@@ -1,4 +1,4 @@
-package io.horizontalsystems.ethereumkit.core.refactoring
+package io.horizontalsystems.ethereumkit.transactionsyncers
 
 import io.horizontalsystems.ethereumkit.api.jsonrpc.models.RpcTransaction
 import io.horizontalsystems.ethereumkit.api.models.AccountState
